# expression: implement vectorized evaluation for `builtinFromUnixTime2ArgSig`