PCP #12106 

### What problem does this PR solve? <!--add issue link with summary if exists-->
Implement vectorized evaluation for `builtinFromUnixTime2ArgSig` at #12106 

### What is changed and how it works?
It gets faster about 33.7%.
Other tiny typo fixes of comments according to `goword` result

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test
```
> GO111MODULE=on go test -check.f TestVectorizedBuiltinTimeFunc
PASS: builtin_time_vec_test.go:202: testEvaluatorSuite.TestVectorizedBuiltinTimeFunc	0.129s
PASS: builtin_time_vec_generated_test.go:172: testEvaluatorSuite.TestVectorizedBuiltinTimeFuncGenerated	0.036s
OK: 2 passed
PASS
ok  	github.com/pingcap/tidb/expression	0.191s
```

 - Benchmark
```
> go test -v -benchmem -bench=BenchmarkVectorizedBuiltinTimeFunc -run=BenchmarkVectorizedBuiltinTimeFunc -args builtinFromUnixTime2ArgSig
goos: darwin
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinTimeFuncGenerated-12    	1000000000	         0.00675 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinFromUnixTime2ArgSig-VecBuiltinFunc-12         	   36835	     31454 ns/op	    6293 B/op	     226 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinFromUnixTime2ArgSig-NonVecBuiltinFunc-12      	   24978	     47419 ns/op	    6293 B/op	     226 allocs/op
PASS
ok  	github.com/pingcap/tidb/expression	3.244s
```

 - Manual test (add detailed scripts or steps below)
```sql
mysql> SELECT FROM_UNIXTIME(1447430881);
+---------------------------+
| FROM_UNIXTIME(1447430881) |
+---------------------------+
| 2015-11-14 00:08:01       |
+---------------------------+
1 row in set (0.00 sec)

mysql> SELECT FROM_UNIXTIME(1447430881) + 0;
+-------------------------------+
| FROM_UNIXTIME(1447430881) + 0 |
+-------------------------------+
|                20151114000801 |
+-------------------------------+
1 row in set (0.00 sec)

mysql> SELECT FROM_UNIXTIME(1447430881, '%Y-%M-%D %H:%i:%S %X');
+---------------------------------------------------+
| FROM_UNIXTIME(1447430881, '%Y-%M-%D %H:%i:%S %X') |
+---------------------------------------------------+
| 2015-November-14th 00:08:01 2015                  |
+---------------------------------------------------+
1 row in set (0.00 sec)

mysql> 
```