@ekalinin 2019-11-10
I'm sorry, but this function was vectorized in #13319